Use io.TextIOWrapper for encoding/decoding
This simplifies things by getting rid of the the hand rolled
encoding/decoding logic in the Reader/Writer objects. By adding
just a few more stubby methods, TextIOWrapper can work with
smart Readers and Writers just fine.

This does change the behavior of http.Reader somewhat in that
it no longer uses the Content-Type charset to decode. Instead
the underlying Reader/Writer objects always work with binary.
If we need to bring back header based decoding at some point
in the future we can figure out an API for that.